--- src/main/java/com/ITP/IFKFbackend/Service/ReportService.java
@@ -1,52 +1,0 @@
-package com.ITP.IFKFbackend.Service;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.util.ResourceUtils;
-
-import com.ITP.IFKFbackend.model.Donations;
-import com.ITP.IFKFbackend.repository.DonationsRepository;
-
-import net.sf.jasperreports.engine.JRException;
-import net.sf.jasperreports.engine.JasperCompileManager;
-import net.sf.jasperreports.engine.JasperExportManager;
-import net.sf.jasperreports.engine.JasperFillManager;
-import net.sf.jasperreports.engine.JasperPrint;
-import net.sf.jasperreports.engine.JasperReport;
-import net.sf.jasperreports.engine.data.JRBeanCollectionDataSource;
-
-@Service
-public class ReportService {
-	
-	@Autowired
-	private DonationsRepository repository;
-	
-	public String getReport() throws FileNotFoundException, JRException {
-		
-		   List<Donations> donations = repository.findAll();
-		//Load a file
-		
-			String path = "C:\\Users\\Senath\\Desktop\\Reports";
-		
-	
-			 File file = ResourceUtils.getFile("classpath:Donations.jrxml");
-			 JasperReport jasperReport = JasperCompileManager.compileReport(file.getAbsolutePath());
-			 JRBeanCollectionDataSource dataSource = new JRBeanCollectionDataSource(donations);
-			 Map<String, Object> parameters = new HashMap<>();
-		     parameters.put("createdBy", "IFKF - Sri Lanka");
-			 JasperPrint jasperPrint = JasperFillManager.fillReport(jasperReport, parameters, dataSource);
-			 JasperExportManager.exportReportToPdfFile(jasperPrint, path+"\\ Donations_Report.pdf");
-			
-			 return "Donations Report Downloaded Successfully in path "+path;
-			
-		
-		
-	}
-
-}
